Add default interval and timeout inputs for monitor task
