+++ src/aiida_workgraph/decorator.py
@@ -382,9 +382,33 @@ class TaskDecoratorCollection:
                 MonitorFunctionTaskFactory,
             )
 
+            sig = inspect.signature(func)
+            forbidden = {"interval", "timeout"}
+            conflicts = forbidden.intersection(sig.parameters)
+            if conflicts:
+                raise ValueError(
+                    f"Function '{func.__name__}' defines parameter(s) {sorted(conflicts)}, "
+                    "which conflict with default 'monitor' arguments."
+                )
+            task_inputs = inputs if inputs is not None else []
+            # Add default interval and timeout
+            task_inputs.extend(
+                [
+                    {
+                        "name": "interval",
+                        "identifier": "workgraph.float",
+                        "property": {"default": 5},
+                    },
+                    {
+                        "name": "timeout",
+                        "identifier": "workgraph.float",
+                        "property": {"default": 3600},
+                    },
+                ]
+            )
             TaskCls = MonitorFunctionTaskFactory.from_function(
                 func=func,
-                inputs=inputs,
+                inputs=task_inputs,
                 outputs=outputs,
             )
